Fix WPA2
This driver works with open networks, but connecting to a network with
WPA2 encryption fails with reason=WRONG_KEY.

Hostapd on the access point says "WPA: WPA IE from (Re)AssocReq did
not match with msg 2/4".

Comparing the RSN information in the association request and in the
"msg 2/4" key reveals that the RSN capabilities are different: 0x0080
versus 0x008c.

For some reason the driver doesn't copy the RSN capabilities into the
association request. I'm not sure where they're supposed to come from
exactly, but the RSN IE in cfg80211_assoc_request seems to have the
correct value, so copy it from there.

Now this driver can connect to networks secured with WPA2.

Signed-off-by: Bitterblue Smith <rtl8821cerfe2@gmail.com>